python/0336 - Above the Rim API/task/app/api/team_schema.py
from marshmallow import Schema, fields, validate


class TeamSchema(Schema):
    short = fields.Str(required=True, validate=[validate.Length(equal=3), str.isupper])
    name = fields.Str(required=True)
