python/0336 - Above the Rim API/requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
Flask
beautifulsoup4
requests
Flask-SQLAlchemy